add readme and comments
