+++ .github/workflows/build.yml
@@ -1,0 +1,28 @@
+name: build
+
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+
+jobs:
+
+  build:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v3
+
+    - name: Set up Go
+      uses: actions/setup-go@v3
+      with:
+        go-version: 1.18
+
+    - name: Install Dependencies
+      run: go get ./...
+
+    - name: Build
+      run: make build
+
+    - name: Test
+      run: make test GOTEST_FLAGS="-count=1"
