+++ README.md
@@ -4,6 +4,18 @@
 #### Source
 A source connector pulls data from bigquery and pushes it to downstream resources via Conduit.
 
+### Implementation
+The connector pulls data from bigquery for a dataset or selected tables of users choice. The connector keeps on
+looking for new insertion happening every minure in any of the table the data is pulled for and syncs it. 
+If the conduit stops or pauses midway the connector will make sure to pull the data which was not pull earlier. 
+for eg,
+table A and table B are synced.
+Pipeline pause after syncing complete table A and table B till index 5.
+On resuming the pipeline - Connector sync data from table B index 6 and would sync table A again if there were no new row added.
+
+Currently only new insertion  i.e row inserted after the last pulled rows are handled. And updataion/deletion is not
+handled by the connector.
+
 ### How to build?
 Run `make build` to build the connector.
 
@@ -14,7 +26,7 @@ in `google_source->source_intergration_test.go`
 The Docker compose file at `test/docker-compose.yml` can be used to run the required resource locally.
 
 ### Known Issues & Limitations
-* Current implementation handles snapshot data.
+* Current implementation handles snapshot and incremental data.
 
 ### Planned work
 - [ ] Handle CDC data
@@ -31,3 +43,4 @@ server to validate configuration and dynamically display configuration options t
 |`srcProjectID`| The Project ID on endpoint|true| false|
 |`srcDatasetID`|The dataset ID to pull data from.|true|false|
 |`srcTableID`|Specify comma separated table IDs. Will pull whole dataset if no Table ID present. |false|false|
+|`dataset_location`|Specify location were dataset exist|true|false|
